gcp-rag-chatbot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2

# Google Cloud SDKs
google-cloud-bigquery==3.25.0
google-cloud-aiplatform==1.66.0
google-auth==2.33.0

# Optional: logging/trace improvements
rich==13.7.1